Update: Update Post services to implement Cloudinary

+++ backend/src/services/postServices.js
@@ -2,12 +2,13 @@ import PostRepository from "../repositories/PostRepository.js";
 import PictureRepository from "../repositories/PictureRepository.js";
 // Importation des classes repository pour post et picture
 import Post from "../classes/Post.js";
-import Picture from "../classes/Picture.js";
-// importations des classes post et picture
+// importation de la classe post
 import prisma from "../prismaClient.js";
 // importation de l'instance du prisma client
 import * as Errors from "../errors/errorsClasses.js";
 // importation de toutes nos classes d'erreurs personnalisées
+import cloudinary from "../../config/cloudinary.js";
+// Importation de notre config cloudinary
 
 export default class PostServices {
   constructor() {
@@ -17,17 +18,28 @@ export default class PostServices {
 
   // POST Création d'un post
   async createPost (data) {
+    let uploadPicture = null;
     try {
       return await prisma.$transaction(async (tx) => {
         if (data.picture) {
-          new Picture(data.picture);
-          const newPicturePost = await this.pictureRepo.createPicture(data.picture, tx);
+          uploadPicture = await cloudinary.uploader.upload(data.picture, {
+            folder: "Posts"
+          });
+          const newPicture = {
+            secureUrl: uploadPicture.secure_url,
+            publicId: uploadPicture.public_id
+          }
+          const newPicturePost = await this.pictureRepo.createPicture(newPicture, tx);
           data.pictureId = newPicturePost.id;
+          delete data.picture;
         }
         new Post(data);
         return await this.postRepo.createPost(data, tx);
       })
     } catch (error) {
+      if (uploadPicture) {
+        await cloudinary.uploader.destroy(uploadPicture.public_id);
+      }
       throw error;
     }
   }
@@ -56,6 +68,8 @@ export default class PostServices {
 
   // PATCH Mise à jour d'un post
   async updatePost (id, data) {
+    // On met une variable en dehors du catch pour être reconnue par catch si elle est modifié
+    let uploadPicture = null;
     try {
       return await prisma.$transaction(async (tx) => {
         // On vérifie que le post existe
@@ -64,30 +78,41 @@ export default class PostServices {
           throw new Errors.NotFoundError("The post doesn't exist.");
         }
   
-  
         // On cree une nouvelle image s'il y en a une et on va suprimmer l'ancienne
-        let oldPicturePostId = null;
         if (data.picture) {
-          new Picture(data.picture);
-          const newPicturePost = await this.pictureRepo.createPicture(data.picture, tx);
-          data.pictureId = newPicturePost.id;
-          if (existingPost.pictureId) {
-            oldPicturePostId = existingPost.pictureId;
+          uploadPicture = await cloudinary.uploader.upload(data.picture, {
+            folder: "Posts"
+          });
+          const newPicture = {
+            secureUrl: uploadPicture.secure_url,
+            publicId: uploadPicture.public_id
           }
+          const newPicturePost = await this.pictureRepo.createPicture(newPicture, tx);
+          // On l'enregistre dans Cloudinary puis on l'enregistre dans la db
+
+          data.pictureId = newPicturePost.id;
+          delete data.picture;
+          // On met l'Id de la nouvelle image dans les données de mise à jour
         }
         const fullPost = {...existingPost, ...data};
         // On fusionne les nouvelles et anciennes données
         new Post(fullPost);
-  
+        // On vérifie le format
         const newPost = await this.postRepo.updatePost(id, data, tx);
 
-        if (oldPicturePostId && oldPicturePostId !== "Id par defaut") {
-          await this.pictureRepo.deletePicture(oldPicturePostId, tx);
+        // On supprime l'ancienne photo s'il y en a une
+        if (uploadPicture && existingPost.picture && existingPost.picture.id !== "Id par defaut") {
+          await cloudinary.uploader.destroy(existingPost.picture.publicId);
+          await this.pictureRepo.deletePicture(existingPost.picture.id, tx);
         }
 
         return newPost;
       })
     } catch (error) {
+      if (uploadPicture) {
+        // On supprime la nouvelle image si le reste de l'opération échoue
+        await cloudinary.uploader.destroy(uploadPicture.public_id);
+      }
       throw error;
     }
   }
@@ -100,7 +125,9 @@ export default class PostServices {
         if (!post) {
           throw new Errors.NotFoundError("The post doesn't exist");
         }
-        if (post.pictureId && post.pictureId !== "Id par defaut") {
+        if (post.pictureId && post.pictureId !== "Id par defaut" && post.picture) {
+          // On supprime l'image de Cloudinary et de notre DB
+          await cloudinary.uploader.destroy(post.picture.publicId);
           await this.pictureRepo.deletePicture(post.pictureId, tx);
         }
         return await this.postRepo.deletePost(id, tx);
